requirements.txt
pyrtlsdr>=0.3.0
